+++ catkin_ws/src/state_estimation/src/ekf_indices.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python
+
+class IDx:
+    # Position
+    Px = 0
+    Py = 1
+    Pz = 2
+    
+    # Velocity
+    Vx = 3
+    Vy = 4
+    Vz = 5
+
+    # Orientation - Quaternion
+    Ow = 6
+    Ox = 7
+    Oy = 8
+    Oz = 9
+
+    # Angular Velocity
+    Ax = 10
+    Ay = 11
+    Az = 12
